move icon map to helpers + remove puppeteer
