+++ jest.config.js
@@ -5,4 +5,5 @@ module.exports = {
     '\\.(css)$': 'identity-obj-proxy',
   },
   collectCoverage: true,
+  preset: 'jest-puppeteer',
 };
